feat(ts): add ts settings (see below, wip) * rename js files to ts files except `forge.config.js` * convert `require()` to `import` * move to `src/` except `forge.config.js` * note: ignore all errors related to ts intentionally due to wip chore(npm-scripts): add typecheck script
