+++ package.json
@@ -7,7 +7,8 @@
     "start": "electron-forge start",
     "package": "electron-forge package",
     "make": "electron-forge make",
-    "publish": "electron-forge publish"
+    "publish": "electron-forge publish",
+    "typecheck": "tsc --noEmit"
   },
   "repository": {
     "type": "git",
